+++ test/animeinfo.js
@@ -3,7 +3,7 @@ const axios = require("axios").default;
 async function animeSearch(query) {
     let run;
     try {
-        const { data } = await axios.get("https://api.jikan.moe/v3/search/anime", {
+        const { data } = await axios.get("https://api.jikan.moe/v4/anime", {
       params: {
         q: query,
         limit: 10